+++ libmpeg/source/mpeg-ts-h264.c
@@ -114,7 +114,7 @@ int mpeg_h264_find_new_access_unit(const uint8_t* data, size_t bytes, int* vcl)
         nal_type = p[n] & 0x1f;
         if (*vcl > 0 && mpeg_h264_is_new_access_unit(p + n, end - p - n))
         {
-            return p - data + n - leading;
+            return (int)(p - data + n - leading);
         }
         else if (nal_type > 0 && nal_type < 6)
         {
